+++ .gitignore
@@ -162,4 +162,8 @@ cython_debug/
 .idea/
 
 data/**/*
-output/**/*
+data/output/**/*
+
+
+.DS_Store
+**/.DS_STORE
