+++ api/src/prompts/__init__.py
@@ -1,0 +1,1 @@
+#https://docs.ankiweb.net/importing/text-files.html
