+++ api/src/utils/image/crop_and_straighten.py
@@ -1,0 +1,50 @@
+import cv2
+import numpy as np
+
+
+def crop_and_straighten(input_path: str):
+    # Read the image
+    image = cv2.imread(input_path)
+
+    # Convert to grayscale
+    gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
+
+    # Apply Gaussian blur
+    blurred = cv2.GaussianBlur(gray, (5, 5), 0)
+
+    # Apply adaptive thresholding
+    thresh = cv2.adaptiveThreshold(blurred, 255, cv2.ADAPTIVE_THRESH_GAUSSIAN_C, cv2.THRESH_BINARY_INV, 11, 2)
+
+    # Find contours
+    contours, _ = cv2.findContours(thresh, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+
+    # Find the largest contour (assuming it's the book)
+    largest_contour = max(contours, key=cv2.contourArea)
+
+    # Get the minimum area rectangle
+    rect = cv2.minAreaRect(largest_contour)
+    box = cv2.boxPoints(rect)
+    box = np.intp(box)
+
+    # Get width and height of the detected rectangle
+    width = int(rect[1][0])
+    height = int(rect[1][1])
+
+    src_pts = box.astype("float32")
+    # Coordinate of the points in box points after the rectangle has been straightened
+    dst_pts = np.array([[0, height - 1],
+                        [0, 0],
+                        [width - 1, 0],
+                        [width - 1, height - 1]], dtype="float32")
+
+    # The perspective transformation matrix
+    M = cv2.getPerspectiveTransform(src_pts, dst_pts)
+
+    # Directly warp the rotated rectangle to get the straightened rectangle
+    warped = cv2.warpPerspective(image, M, (width, height))
+
+    # Add a small padding
+    padding = 5
+    warped = cv2.copyMakeBorder(warped, padding, padding, padding, padding, cv2.BORDER_CONSTANT, value=[255, 255, 255])
+
+    return warped
