+++ api/src/utils/prompt.py
@@ -1,0 +1,2 @@
+# GENERATE PROMPTS TO BE SENT TO MESSAGES TO OPENAPI
+
