--- snap-to-anki/prompts/__init__.py
@@ -1,1 +1,0 @@
-#https://docs.ankiweb.net/importing/text-files.html
