--- snap-to-anki/utils/prompt.py
@@ -1,2 +1,0 @@
-# GENERATE PROMPTS TO BE SENT TO MESSAGES TO OPENAPI
-
